Mission segment separation
Mission segments are now in separate files and registered using submodels
