+++ src/fastga/models/performances/mission/constants.py
@@ -1,0 +1,19 @@
+"""Constants for the geometry submodels."""
+#  This file is part of FAST : A framework for rapid Overall Aircraft Design
+#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  FAST is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+SUBMODEL_TAXI = "submodel.performances.mission.taxi"
+SUBMODEL_CLIMB = "submodel.performances.mission.climb"
+SUBMODEL_CRUISE = "submodel.performances.mission.cruise"
+SUBMODEL_DESCENT = "submodel.performances.mission.descent"
+SUBMODEL_RESERVES = "submodel.performances.mission.reserves"
